# backend/bookmyseat/src/main/java/com/example/bookmyseat/dto/ShowDto.java
package com.example.bookmyseat.dto;

import lombok.Builder;
import lombok.Data;
import java.time.LocalDateTime;
import java.util.UUID;

@Data
@Builder
public class ShowDto {
    private UUID id;
    private UUID movieId;
    private String movieTitle;
    private UUID theaterId;
    private String theaterName;
    private String screenName;
    private LocalDateTime startTime;
    private LocalDateTime endTime;
}